refactor: Modularize and clean up locator generation and XML parsing components
- Removed `generate-all-locators.ts` and `source-parsing.ts`, integrating their functionality into more focused modules (`xml-parsing`, `constants`, `types`) for improved maintainability.
- Consolidated shared types and constants for streamlined usage
