+++ src/locators/locator-generation.ts
@@ -1,30 +1,9 @@
 /**
- * Generate multiple locator strategies for an element
- *
- * Based on: https://github.com/appium/appium-mcp/blob/main/src/locators/locator-generation.ts
+ * Locator strategy generation for mobile elements
  */
 
-import type { JSONElement, UniquenessResult } from './source-parsing';
-import { checkXPathUniqueness, evaluateXPath, isAttributeUnique } from './source-parsing';
-
-/**
- * Context for locator generation - carries DOM state
- */
-export interface LocatorContext {
-  sourceXML: string;
-  parsedDOM: Document | null;
-  isAndroid: boolean;
-}
-
-export type LocatorStrategy =
-  | 'accessibility-id'
-  | 'id'
-  | 'class-name'
-  | 'xpath'
-  | 'predicate-string'
-  | 'class-chain'
-  | 'uiautomator'
-  | 'text';
+import type { JSONElement, LocatorStrategy, LocatorContext, UniquenessResult } from './types';
+import { checkXPathUniqueness, evaluateXPath, isAttributeUnique } from './xml-parsing';
 
 /**
  * Check if a string value is valid for use in a locator
@@ -42,7 +21,6 @@ function escapeText(text: string): string {
 
 /**
  * Escape value for use in XPath expressions
- * Handles values containing both single and double quotes using concat()
  */
 function escapeXPathValue(value: string): string {
   if (!value.includes("'")) {
@@ -51,7 +29,6 @@ function escapeXPathValue(value: string): string {
   if (!value.includes('"')) {
     return `"${value}"`;
   }
-  // Value contains both quotes - use concat()
   const parts: string[] = [];
   let current = '';
   for (const char of value) {
@@ -68,75 +45,37 @@ function escapeXPathValue(value: string): string {
 }
 
 /**
- * Wrap non-unique XPath with index: //button[@text="OK"] → (//button[@text="OK"])[2]
+ * Wrap non-unique XPath with index
  */
 function generateIndexedXPath(baseXPath: string, index: number): string {
   return `(${baseXPath})[${index}]`;
 }
 
 /**
- * Add .instance(n) for UiAutomator: new UiSelector().text("OK") → new UiSelector().text("OK").instance(1)
- * Note: UiAutomator instance() is 0-based
+ * Add .instance(n) for UiAutomator (0-based)
  */
 function generateIndexedUiAutomator(baseSelector: string, index: number): string {
-  // UiAutomator uses 0-based indexing, XPath uses 1-based
   return `${baseSelector}.instance(${index - 1})`;
 }
 
 /**
- * Add XPath locator with uniqueness checking and fallbacks
- * Handles indexed XPath and hierarchical XPath generation
+ * Check uniqueness, falling back to regex if no DOM available
  */
-function addXPathLocator(
-  results: [LocatorStrategy, string][],
-  xpath: string,
+function checkUniqueness(
   ctx: LocatorContext,
+  xpath: string,
   targetNode?: Node,
-): void {
-  const uniqueness = checkUniqueness(ctx, xpath, targetNode);
-  if (uniqueness.isUnique) {
-    results.push(['xpath', xpath]);
-  } else if (uniqueness.index) {
-    results.push(['xpath', generateIndexedXPath(xpath, uniqueness.index)]);
-  } else {
-    // XPath not unique and no index - try hierarchical XPath
-    if (targetNode && ctx.parsedDOM) {
-      const hierarchical = buildHierarchicalXPath(ctx, targetNode as Element);
-      if (hierarchical) {
-        results.push(['xpath', hierarchical]);
-      }
-    }
-    // Still add the non-unique xpath as last resort
-    results.push(['xpath', xpath]);
+): UniquenessResult {
+  if (ctx.parsedDOM) {
+    return checkXPathUniqueness(ctx.parsedDOM, xpath, targetNode);
   }
-}
 
-/**
- * Check if element is within UiAutomator scope
- * UiAutomator only works within /hierarchy/*[last()] on Android
- * Elements outside this scope (e.g., in status bar) need XPath fallback
- */
-function isInUiAutomatorScope(element: JSONElement, doc: Document | null): boolean {
-  if (!doc) return true; // Assume in scope if no DOM available
-
-  // Get hierarchy children
-  const hierarchyNodes = evaluateXPath(doc, '/hierarchy/*');
-  if (hierarchyNodes.length === 0) return true; // No hierarchy, assume in scope
-
-  // The last child of /hierarchy is the app content
-  // UiAutomator can only interact with elements in this subtree
-  const lastIndex = hierarchyNodes.length;
-
-  // Element path starts with index into hierarchy children
-  // e.g., path "0.1.2" means child 0 of hierarchy
-  const pathParts = element.path.split('.');
-  if (pathParts.length === 0 || pathParts[0] === '') return true;
-
-  const firstIndex = parseInt(pathParts[0], 10);
-
-  // Element is in scope if it's in the last hierarchy child
-  // (index is 0-based, so lastIndex - 1 is the last child)
-  return firstIndex === lastIndex - 1;
+  const match = xpath.match(/\/\/\*\[@([^=]+)="([^"]+)"\]/);
+  if (match) {
+    const [, attr, value] = match;
+    return { isUnique: isAttributeUnique(ctx.sourceXML, attr, value) };
+  }
+  return { isUnique: false };
 }
 
 /**
@@ -161,13 +100,29 @@ function getSiblingIndex(element: Element): number {
 }
 
 /**
- * Find unique attribute for element in XPath format
- * @returns Attribute selector like @resource-id="value" or null if none unique
+ * Count siblings with same tag name
  */
-function findUniqueAttribute(
-  element: Element,
-  ctx: LocatorContext,
-): string | null {
+function countSiblings(element: Element): number {
+  const parent = element.parentNode;
+  if (!parent) return 1;
+
+  const tagName = element.nodeName;
+  let count = 0;
+
+  for (let i = 0; i < parent.childNodes.length; i++) {
+    const child = parent.childNodes.item(i);
+    if (child?.nodeType === 1 && child.nodeName === tagName) {
+      count++;
+    }
+  }
+
+  return count;
+}
+
+/**
+ * Find unique attribute for element in XPath format
+ */
+function findUniqueAttribute(element: Element, ctx: LocatorContext): string | null {
   const attrs = ctx.isAndroid
     ? ['resource-id', 'content-desc', 'text']
     : ['name', 'label', 'value'];
@@ -191,7 +146,6 @@ function findUniqueAttribute(
 
 /**
  * Build hierarchical XPath by traversing up the DOM tree
- * Creates paths like: //parent[@id="x"]/child[2]/target
  */
 function buildHierarchicalXPath(
   ctx: LocatorContext,
@@ -209,11 +163,9 @@ function buildHierarchicalXPath(
     const uniqueAttr = findUniqueAttribute(current, ctx);
 
     if (uniqueAttr) {
-      // Found unique attribute - we can stop here
       pathParts.unshift(`//${tagName}[${uniqueAttr}]`);
       break;
     } else {
-      // Use sibling index
       const siblingIndex = getSiblingIndex(current);
       const siblingCount = countSiblings(current);
 
@@ -225,20 +177,17 @@ function buildHierarchicalXPath(
     }
 
     const parent = current.parentNode as Element | null;
-    // Stop at non-element nodes
-    current = (parent && parent.nodeType === 1) ? parent : null;
+    current = parent && parent.nodeType === 1 ? parent : null;
     depth++;
   }
 
   if (pathParts.length === 0) return null;
 
-  // Join parts - first part already has // prefix if it found unique attr
   let result = pathParts[0];
   for (let i = 1; i < pathParts.length; i++) {
     result += '/' + pathParts[i];
   }
 
-  // If no unique attr found, prefix with //
   if (!result.startsWith('//')) {
     result = '//' + result;
   }
@@ -247,147 +196,45 @@ function buildHierarchicalXPath(
 }
 
 /**
- * Count siblings with same tag name
+ * Add XPath locator with uniqueness checking and fallbacks
  */
-function countSiblings(element: Element): number {
-  const parent = element.parentNode;
-  if (!parent) return 1;
-
-  const tagName = element.nodeName;
-  let count = 0;
-
-  for (let i = 0; i < parent.childNodes.length; i++) {
-    const child = parent.childNodes.item(i);
-    if (child?.nodeType === 1 && child.nodeName === tagName) {
-      count++;
-    }
-  }
-
-  return count;
-}
-
-/**
- * Get simple locators based on single attributes
- * These are preferred because they're most stable and readable
- */
-function getSimpleSuggestedLocators(
-  element: JSONElement,
-  ctx: LocatorContext,
-  automationName: string,
-  targetNode?: Node,
-): [LocatorStrategy, string][] {
-  const results: [LocatorStrategy, string][] = [];
-  const isAndroid = automationName.toLowerCase().includes('uiautomator');
-  const attrs = element.attributes;
-
-  // Check if element is in UiAutomator scope (Android only)
-  const inUiAutomatorScope = isAndroid ? isInUiAutomatorScope(element, ctx.parsedDOM) : true;
-
-  if (isAndroid) {
-    // Android simple locators
-
-    // 1. Resource ID (most stable)
-    const resourceId = attrs['resource-id'];
-    if (isValidValue(resourceId)) {
-      const xpath = `//*[@resource-id="${resourceId}"]`;
-      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
-
-      if (uniqueness.isUnique && inUiAutomatorScope) {
-        results.push(['id', `android=new UiSelector().resourceId("${resourceId}")`]);
-      } else if (uniqueness.index && inUiAutomatorScope) {
-        // Generate indexed UiAutomator selector
-        const base = `android=new UiSelector().resourceId("${resourceId}")`;
-        results.push(['id', generateIndexedUiAutomator(base, uniqueness.index)]);
-      }
-    }
-
-    // 2. Content Description (accessibility)
-    const contentDesc = attrs['content-desc'];
-    if (isValidValue(contentDesc)) {
-      const xpath = `//*[@content-desc="${contentDesc}"]`;
-      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
-
-      if (uniqueness.isUnique) {
-        results.push(['accessibility-id', `~${contentDesc}`]);
-      }
-      // Note: accessibility-id (~) doesn't support indexing, skip if not unique
-    }
-
-    // 3. Text (visible text)
-    const text = attrs.text;
-    if (isValidValue(text) && text.length < 100) {
-      const xpath = `//*[@text="${escapeText(text)}"]`;
-      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
-
-      if (uniqueness.isUnique && inUiAutomatorScope) {
-        results.push(['text', `android=new UiSelector().text("${escapeText(text)}")`]);
-      } else if (uniqueness.index && inUiAutomatorScope) {
-        const base = `android=new UiSelector().text("${escapeText(text)}")`;
-        results.push(['text', generateIndexedUiAutomator(base, uniqueness.index)]);
-      }
-    }
-  } else {
-    // iOS simple locators
-
-    // 1. Accessibility ID (name attribute)
-    const name = attrs.name;
-    if (isValidValue(name)) {
-      const xpath = `//*[@name="${name}"]`;
-      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
-
-      if (uniqueness.isUnique) {
-        results.push(['accessibility-id', `~${name}`]);
-      }
-      // Note: accessibility-id (~) doesn't support indexing
-    }
-
-    // 2. Label (visible text, often same as name)
-    const label = attrs.label;
-    if (isValidValue(label) && label !== attrs.name) {
-      const xpath = `//*[@label="${escapeText(label)}"]`;
-      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
-
-      if (uniqueness.isUnique) {
-        results.push(['predicate-string', `-ios predicate string:label == "${escapeText(label)}"`]);
-      }
-      // Note: predicate-string doesn't have native indexing - XPath fallback will handle
-    }
-
-    // 3. Value
-    const value = attrs.value;
-    if (isValidValue(value)) {
-      const xpath = `//*[@value="${escapeText(value)}"]`;
-      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
-
-      if (uniqueness.isUnique) {
-        results.push(['predicate-string', `-ios predicate string:value == "${escapeText(value)}"`]);
-      }
-    }
-  }
-
-  return results;
-}
-
-/**
- * Check uniqueness, falling back to regex if no DOM available
- */
-function checkUniqueness(
-  ctx: LocatorContext,
+function addXPathLocator(
+  results: [LocatorStrategy, string][],
   xpath: string,
+  ctx: LocatorContext,
   targetNode?: Node,
-): UniquenessResult {
-  if (ctx.parsedDOM) {
-    return checkXPathUniqueness(ctx.parsedDOM, xpath, targetNode);
+): void {
+  const uniqueness = checkUniqueness(ctx, xpath, targetNode);
+  if (uniqueness.isUnique) {
+    results.push(['xpath', xpath]);
+  } else if (uniqueness.index) {
+    results.push(['xpath', generateIndexedXPath(xpath, uniqueness.index)]);
+  } else {
+    if (targetNode && ctx.parsedDOM) {
+      const hierarchical = buildHierarchicalXPath(ctx, targetNode as Element);
+      if (hierarchical) {
+        results.push(['xpath', hierarchical]);
+      }
+    }
+    results.push(['xpath', xpath]);
   }
+}
 
-  // Fallback to regex-based check (no index info available)
-  // Extract attribute and value from xpath pattern //*[@attr="value"]
-  const match = xpath.match(/\/\/\*\[@([^=]+)="([^"]+)"\]/);
-  if (match) {
-    const [, attr, value] = match;
-    return { isUnique: isAttributeUnique(ctx.sourceXML, attr, value) };
-  }
-  return { isUnique: false };
+/**
+ * Check if element is within UiAutomator scope
+ */
+function isInUiAutomatorScope(element: JSONElement, doc: Document | null): boolean {
+  if (!doc) return true;
+
+  const hierarchyNodes = evaluateXPath(doc, '/hierarchy/*');
+  if (hierarchyNodes.length === 0) return true;
+
+  const lastIndex = hierarchyNodes.length;
+  const pathParts = element.path.split('.');
+  if (pathParts.length === 0 || pathParts[0] === '') return true;
+
+  const firstIndex = parseInt(pathParts[0], 10);
+  return firstIndex === lastIndex - 1;
 }
 
 /**
@@ -397,7 +244,6 @@ function buildUiAutomatorSelector(element: JSONElement): string | null {
   const attrs = element.attributes;
   const parts: string[] = [];
 
-  // Build selector with available attributes
   if (isValidValue(attrs['resource-id'])) {
     parts.push(`resourceId("${attrs['resource-id']}")`);
   }
@@ -441,7 +287,6 @@ function buildPredicateString(element: JSONElement): string | null {
 
   if (conditions.length === 0) return null;
 
-  // Use AND for multiple conditions
   return `-ios predicate string:${conditions.join(' AND ')}`;
 }
 
@@ -452,12 +297,10 @@ function buildClassChain(element: JSONElement): string | null {
   const attrs = element.attributes;
   const tagName = element.tagName;
 
-  // Simple class chain with type
   if (!tagName.startsWith('XCUI')) return null;
 
   let selector = `**/${tagName}`;
 
-  // Add label predicate if available
   if (isValidValue(attrs.label)) {
     selector += `[\`label == "${escapeText(attrs.label!)}"\`]`;
   } else if (isValidValue(attrs.name)) {
@@ -476,7 +319,6 @@ function buildXPath(element: JSONElement, sourceXML: string, isAndroid: boolean)
   const conditions: string[] = [];
 
   if (isAndroid) {
-    // Android XPath attributes
     if (isValidValue(attrs['resource-id'])) {
       conditions.push(`@resource-id="${attrs['resource-id']}"`);
     }
@@ -487,7 +329,6 @@ function buildXPath(element: JSONElement, sourceXML: string, isAndroid: boolean)
       conditions.push(`@text="${escapeText(attrs.text!)}"`);
     }
   } else {
-    // iOS XPath attributes
     if (isValidValue(attrs.name)) {
       conditions.push(`@name="${attrs.name}"`);
     }
@@ -499,14 +340,102 @@ function buildXPath(element: JSONElement, sourceXML: string, isAndroid: boolean)
     }
   }
 
-  // Build XPath
   if (conditions.length === 0) {
-    // Fallback: just the tag
     return `//${tagName}`;
   }
 
-  // Combine conditions with 'and'
   return `//${tagName}[${conditions.join(' and ')}]`;
+}
+
+/**
+ * Get simple locators based on single attributes
+ */
+function getSimpleSuggestedLocators(
+  element: JSONElement,
+  ctx: LocatorContext,
+  automationName: string,
+  targetNode?: Node,
+): [LocatorStrategy, string][] {
+  const results: [LocatorStrategy, string][] = [];
+  const isAndroid = automationName.toLowerCase().includes('uiautomator');
+  const attrs = element.attributes;
+  const inUiAutomatorScope = isAndroid ? isInUiAutomatorScope(element, ctx.parsedDOM) : true;
+
+  if (isAndroid) {
+    // Resource ID
+    const resourceId = attrs['resource-id'];
+    if (isValidValue(resourceId)) {
+      const xpath = `//*[@resource-id="${resourceId}"]`;
+      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
+
+      if (uniqueness.isUnique && inUiAutomatorScope) {
+        results.push(['id', `android=new UiSelector().resourceId("${resourceId}")`]);
+      } else if (uniqueness.index && inUiAutomatorScope) {
+        const base = `android=new UiSelector().resourceId("${resourceId}")`;
+        results.push(['id', generateIndexedUiAutomator(base, uniqueness.index)]);
+      }
+    }
+
+    // Content Description
+    const contentDesc = attrs['content-desc'];
+    if (isValidValue(contentDesc)) {
+      const xpath = `//*[@content-desc="${contentDesc}"]`;
+      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
+
+      if (uniqueness.isUnique) {
+        results.push(['accessibility-id', `~${contentDesc}`]);
+      }
+    }
+
+    // Text
+    const text = attrs.text;
+    if (isValidValue(text) && text.length < 100) {
+      const xpath = `//*[@text="${escapeText(text)}"]`;
+      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
+
+      if (uniqueness.isUnique && inUiAutomatorScope) {
+        results.push(['text', `android=new UiSelector().text("${escapeText(text)}")`]);
+      } else if (uniqueness.index && inUiAutomatorScope) {
+        const base = `android=new UiSelector().text("${escapeText(text)}")`;
+        results.push(['text', generateIndexedUiAutomator(base, uniqueness.index)]);
+      }
+    }
+  } else {
+    // iOS: Accessibility ID (name)
+    const name = attrs.name;
+    if (isValidValue(name)) {
+      const xpath = `//*[@name="${name}"]`;
+      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
+
+      if (uniqueness.isUnique) {
+        results.push(['accessibility-id', `~${name}`]);
+      }
+    }
+
+    // iOS: Label
+    const label = attrs.label;
+    if (isValidValue(label) && label !== attrs.name) {
+      const xpath = `//*[@label="${escapeText(label)}"]`;
+      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
+
+      if (uniqueness.isUnique) {
+        results.push(['predicate-string', `-ios predicate string:label == "${escapeText(label)}"`]);
+      }
+    }
+
+    // iOS: Value
+    const value = attrs.value;
+    if (isValidValue(value)) {
+      const xpath = `//*[@value="${escapeText(value)}"]`;
+      const uniqueness = checkUniqueness(ctx, xpath, targetNode);
+
+      if (uniqueness.isUnique) {
+        results.push(['predicate-string', `-ios predicate string:value == "${escapeText(value)}"`]);
+      }
+    }
+  }
+
+  return results;
 }
 
 /**
@@ -520,14 +449,9 @@ function getComplexSuggestedLocators(
 ): [LocatorStrategy, string][] {
   const results: [LocatorStrategy, string][] = [];
   const isAndroid = automationName.toLowerCase().includes('uiautomator');
-
-  // Check if element is in UiAutomator scope (Android only)
   const inUiAutomatorScope = isAndroid ? isInUiAutomatorScope(element, ctx.parsedDOM) : true;
 
   if (isAndroid) {
-    // Android complex locators
-
-    // UiAutomator with multiple attributes (only if in scope)
     if (inUiAutomatorScope) {
       const uiAutomator = buildUiAutomatorSelector(element);
       if (uiAutomator) {
@@ -535,13 +459,11 @@ function getComplexSuggestedLocators(
       }
     }
 
-    // XPath with uniqueness checking
     const xpath = buildXPath(element, ctx.sourceXML, true);
     if (xpath) {
       addXPathLocator(results, xpath, ctx, targetNode);
     }
 
-    // Class name (least specific, only if in scope)
     if (inUiAutomatorScope && isValidValue(element.attributes.class)) {
       results.push([
         'class-name',
@@ -549,27 +471,21 @@ function getComplexSuggestedLocators(
       ]);
     }
   } else {
-    // iOS complex locators
-
-    // Predicate string with multiple conditions
     const predicate = buildPredicateString(element);
     if (predicate) {
       results.push(['predicate-string', predicate]);
     }
 
-    // Class chain
     const classChain = buildClassChain(element);
     if (classChain) {
       results.push(['class-chain', classChain]);
     }
 
-    // XPath with uniqueness checking
     const xpath = buildXPath(element, ctx.sourceXML, false);
     if (xpath) {
       addXPathLocator(results, xpath, ctx, targetNode);
     }
 
-    // Class name (least specific)
     const type = element.tagName;
     if (type.startsWith('XCUIElementType')) {
       results.push(['class-name', `-ios class chain:**/${type}`]);
@@ -581,17 +497,6 @@ function getComplexSuggestedLocators(
 
 /**
  * Get all suggested locators for an element
- * Returns array of [strategy, value] tuples ordered by priority
- *
- * Priority order:
- * Android: id > accessibility-id > text > xpath > uiautomator > class-name
- * iOS: accessibility-id > predicate-string > class-chain > xpath > class-name
- *
- * @param element - JSONElement from parsed page source
- * @param sourceXML - Original XML string (used for regex fallback)
- * @param automationName - Automation driver name (uiautomator2/xcuitest)
- * @param ctx - Optional locator context with cached DOM (for performance)
- * @param targetNode - Optional DOM node for this element (for indexed locators)
  */
 export function getSuggestedLocators(
   element: JSONElement,
@@ -600,20 +505,15 @@ export function getSuggestedLocators(
   ctx?: LocatorContext,
   targetNode?: Node,
 ): [LocatorStrategy, string][] {
-  // Create context if not provided
   const locatorCtx = ctx ?? {
     sourceXML,
     parsedDOM: null,
     isAndroid: automationName.toLowerCase().includes('uiautomator'),
   };
 
-  // Get simple (single attribute) locators first
   const simpleLocators = getSimpleSuggestedLocators(element, locatorCtx, automationName, targetNode);
-
-  // Get complex (combination) locators
   const complexLocators = getComplexSuggestedLocators(element, locatorCtx, automationName, targetNode);
 
-  // Combine, removing duplicates (keep first occurrence)
   const seen = new Set<string>();
   const results: [LocatorStrategy, string][] = [];
 
@@ -645,7 +545,6 @@ export function getBestLocator(
 export function locatorsToObject(locators: [LocatorStrategy, string][]): Record<string, string> {
   const result: Record<string, string> = {};
   for (const [strategy, value] of locators) {
-    // Use first locator for each strategy
     if (!result[strategy]) {
       result[strategy] = value;
     }
